requirements.txt
# Requirements for Python Dependencies

PyPDF2>=3.0.0
python-pptx>=0.6.21
python-docx>=0.8.11
